Use jlong to store pointer for 64-bit Linux support
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@1227 06c8f799-c779-4783-8af7-179afedb67c8

+++ pccomms/lejos/pc/comm/NXTCommFantom.java
@@ -6,10 +6,10 @@ public class NXTCommFantom implements NXTComm {
 	private NXTInfo nxtInfo;
 	
 	public native String[] jfantom_find();
-	public native int jfantom_open(String nxt);
-	public native void jfantom_close(int nxt);
-	public native void jfantom_send_data(int nxt, byte [] message, int len, int replyLen);
-	public native byte[] jfantom_read_data(int nxt, int len);
+	public native long jfantom_open(String nxt);
+	public native void jfantom_close(long nxt);
+	public native void jfantom_send_data(long nxt, byte [] message, int len, int replyLen);
+	public native byte[] jfantom_read_data(long nxt, int len);
 	
 	public NXTInfo[] search(String name, int protocol) {
 		String[] nxtNames = jfantom_find();
